fix(main/rust): always export __emutls_get_address
in version script when linking

Also update test coverage and check for unused variables

+++ packages/rust/tests/0001-cargo-build.sh
@@ -1,15 +1,26 @@
 #!/bin/bash
-set -e
+set -e -u
 
-pkg install -y file ndk-multilib wasm-component-ld
-#pkg install -y rust rust-std-*
+echo_and_run() {
+	echo "> $*"
+	bash -c "$*"
+}
+
+echo_and_run pkg install -y file llvm ndk-multilib wasm-component-ld
+#echo_and_run pkg install -y rust rust-std-*
+
+command -v cargo || (echo "cargo is not installed" && exit 1)
+cargo -V
+command -v rustc || (echo "rustc is not installed" && exit 1)
+rustc -V
 
 export RUSTC_LOG="rustc_codegen_ssa::back::link=info"
 echo "RUSTC_LOG=$RUSTC_LOG"
 
 tmpdir=$(mktemp -d)
 pushd "$tmpdir"
-cargo new rusthello
+
+echo_and_run cargo new rusthello
 pushd rusthello
 mkdir -p .cargo
 cat <<- EOL > .cargo/config.toml
@@ -22,30 +33,47 @@ rustflags = "-Clink-arg=--target=i686-linux-android"
 [target.x86_64-linux-android]
 rustflags = "-Clink-arg=--target=x86_64-linux-android"
 EOL
-cargo build
-cargo build --target aarch64-linux-android
-cargo build --target armv7-linux-androideabi
-cargo build --target i686-linux-android
-cargo build --target x86_64-linux-android
-cargo build --target wasm32-unknown-unknown
-cargo build --target wasm32-wasip1
-cargo build --target wasm32-wasip2
-cargo build --release
-cargo build --target aarch64-linux-android --release
-cargo build --target armv7-linux-androideabi --release
-cargo build --target i686-linux-android --release
-cargo build --target x86_64-linux-android --release
-cargo build --target wasm32-unknown-unknown --release
-cargo build --target wasm32-wasip1 --release
-cargo build --target wasm32-wasip2 --release
+echo_and_run cargo build
+echo_and_run cargo build --target aarch64-linux-android
+echo_and_run cargo build --target armv7-linux-androideabi
+echo_and_run cargo build --target i686-linux-android
+echo_and_run cargo build --target x86_64-linux-android
+echo_and_run cargo build --target wasm32-unknown-unknown
+echo_and_run cargo build --target wasm32-wasip1
+echo_and_run cargo build --target wasm32-wasip2
+echo_and_run cargo build --release
+echo_and_run cargo build --target aarch64-linux-android --release
+echo_and_run cargo build --target armv7-linux-androideabi --release
+echo_and_run cargo build --target i686-linux-android --release
+echo_and_run cargo build --target x86_64-linux-android --release
+echo_and_run cargo build --target wasm32-unknown-unknown --release
+echo_and_run cargo build --target wasm32-wasip1 --release
+echo_and_run cargo build --target wasm32-wasip2 --release
+echo_and_run cargo run
+echo_and_run file target/*/rusthello
+echo_and_run file target/*/*/rusthello
+echo_and_run file target/*/*/rusthello.wasm
+echo_and_run llvm-readelf -d target/*/rusthello
+if ! echo_and_run "llvm-readelf -s target/*/rusthello | grep __emutls_get_address | grep WEAK"; then
+	llvm-readelf -s target/*/rusthello | grep __emutls_get_address
+	echo "ERROR: this should be WEAK" 1>&2
+	exit 1
+fi
 popd
-file rusthello/target/*/rusthello
-file rusthello/target/*/*/rusthello
-file rusthello/target/*/*/rusthello.wasm
+
+echo_and_run cargo new rusthello-lib --lib
+pushd rusthello-lib
+echo >> Cargo.toml
+echo '[lib]' >> Cargo.toml
+echo 'crate-type = ["cdylib"]' >> Cargo.toml
+echo_and_run cargo build
+echo_and_run llvm-readelf -d target/*/deps/librusthello_lib.so
+if ! echo_and_run "llvm-readelf -s target/*/deps/librusthello_lib.so | grep __emutls_get_address | grep WEAK"; then
+	llvm-readelf -s target/*/deps/librusthello_lib.so | grep __emutls_get_address
+	echo "ERROR: this should be WEAK" 1>&2
+	exit 1
+fi
+popd
+
 popd
 rm -fr "$tmpdir"
-
-command -v cargo
-cargo -V
-command -v rustc
-rustc -V
